Fixed Memory bug
Also updated addresses to use uints

+++ PashLang/PASM/Handlers/set_PAR.cs
@@ -25,8 +25,8 @@ namespace PASM.Handlers
         int cpy_ptr;
 
         //Ranges we want to copy with
-        int startingRange;
-        int endingRange;
+        uint startingRange;
+        uint endingRange;
 
         public set_PAR(string[] args, Engine inst) : base(inst)
         {
@@ -34,8 +34,8 @@ namespace PASM.Handlers
 
             cpy_isMethodPtr = isMethodPointer(args[3], out cpy_ptr);
 
-            startingRange = Converter.ParseStringToInt(args[4]);
-            endingRange = Converter.ParseStringToInt(args[5]);
+            startingRange = (uint)Converter.ParseStringToInt(args[4]);
+            endingRange = (uint)Converter.ParseStringToInt(args[5]);
 
         }
